Removed edge on click because it was to slow. Moved the showInfo to gui controller.
Edge onClick will be implemented once dynamic loading is added because that makes it possible to use the layers that are introduced there.
